% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19SeverityPredictionwWithoutCV.R
\docType{package}
\name{Covid19SeverityPredictionwWithoutCV}
\alias{Covid19SeverityPredictionwWithoutCV}
\alias{Covid19SeverityPredictionwWithoutCV-package}
\title{Covid19SeverityPredictionwWithoutCV}
\description{
Covid19SeverityPredictionwWithoutCV
}
